Move APIv1/HTTP to package org.projectnessie.client.http.v1api
