--- clients/client/src/main/java/org/projectnessie/client/http/HttpCreateReference.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (C) 2020 Dremio
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.projectnessie.client.http;
-
-import org.projectnessie.client.api.CreateReferenceBuilder;
-import org.projectnessie.error.NessieConflictException;
-import org.projectnessie.error.NessieNotFoundException;
-import org.projectnessie.model.Reference;
-
-final class HttpCreateReference extends BaseHttpRequest implements CreateReferenceBuilder {
-
-  private Reference reference;
-
-  HttpCreateReference(NessieHttpClient client) {
-    super(client);
-  }
-
-  @Override
-  public CreateReferenceBuilder reference(Reference reference) {
-    this.reference = reference;
-    return this;
-  }
-
-  @Override
-  public Reference submit() throws NessieNotFoundException, NessieConflictException {
-    return client.getTreeApi().createReference(reference);
-  }
-}
