+++ clients/client/src/main/java/org/projectnessie/client/http/v1api/HttpGetCommitLog.java
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.client.http;
+package org.projectnessie.client.http.v1api;
 
 import org.projectnessie.api.params.CommitLogParams;
 import org.projectnessie.client.api.GetCommitLogBuilder;
+import org.projectnessie.client.http.NessieHttpClient;
 import org.projectnessie.error.NessieNotFoundException;
 import org.projectnessie.model.LogResponse;
