+++ clients/client/src/main/java/org/projectnessie/client/http/v1api/HttpGetContents.java
@@ -13,12 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.client.http;
+package org.projectnessie.client.http.v1api;
 
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import org.projectnessie.client.api.GetContentsBuilder;
+import org.projectnessie.client.http.NessieHttpClient;
 import org.projectnessie.error.NessieNotFoundException;
 import org.projectnessie.model.Contents;
 import org.projectnessie.model.ContentsKey;
