+++ clients/client/src/main/java/org/projectnessie/client/http/v1api/HttpGetReference.java
@@ -13,20 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.projectnessie.client.http;
+package org.projectnessie.client.http.v1api;
 
-import java.util.List;
-import org.projectnessie.client.api.GetAllReferencesBuilder;
+import org.projectnessie.client.api.GetReferenceBuilder;
+import org.projectnessie.client.http.NessieHttpClient;
+import org.projectnessie.error.NessieNotFoundException;
 import org.projectnessie.model.Reference;
 
-final class HttpGetAllReferences extends BaseHttpRequest implements GetAllReferencesBuilder {
+final class HttpGetReference extends BaseHttpRequest implements GetReferenceBuilder {
 
-  HttpGetAllReferences(NessieHttpClient client) {
+  private String refName;
+
+  HttpGetReference(NessieHttpClient client) {
     super(client);
   }
 
   @Override
-  public List<Reference> submit() {
-    return client.getTreeApi().getAllReferences();
+  public GetReferenceBuilder refName(String refName) {
+    this.refName = refName;
+    return this;
+  }
+
+  @Override
+  public Reference submit() throws NessieNotFoundException {
+    return client.getTreeApi().getReferenceByName(refName);
   }
 }
